update: フォームとロジックを分離
