+++ src/components/pages/Login/index.tsx
@@ -1,14 +1,14 @@
 import PageLayout from '@/components/bases/PageLayout';
-import LoginForm from '@/features/Auth/component/LoginForm';
+import Login from '@/features/Auth/component/Login';
 
-export default function Login() {
+export default function LoginPage() {
   return (
     <PageLayout title='ログイン'>
       <div className='flex h-full flex-col items-center justify-center'>
         <p>
           プレースホルダーのメアドとパスワードでログインできます。会員登録もできます。
         </p>
-        <LoginForm loginSuccessRedirect='/menu' />
+        <Login loginSuccessRedirect='/menu' />
       </div>
     </PageLayout>
   );
